refactor(cef): split host reducer.rs into mod.rs + tests.rs (task #182 PR-F-1)
Same mechanical extraction as launcher PR-A. Host reducer was 2307
lines (33% tests). Split:

- agentmux-cef/src/reducer.rs → agentmux-cef/src/reducer/mod.rs (1537 lines, prod)
- New: agentmux-cef/src/reducer/tests.rs (770 lines, all tests)

mod.rs ends with #[cfg(test)] mod tests;. tests.rs starts with
use super::*; (de-indented from inline mod tests block).

This is the first bite of PR-F (host reducer split). Subsequent
PRs will extract per-Phase-H-domain handlers (panes, browsers, drag,
pool, quit, top-level) following the same pattern as the launcher
reducer split.

63 host tests pass.
